requirements.txt
h5netcdf
h5py
netCDF4
numpy
xarray
xarray-datatree
